js: moved renderEnvControls() to js/topbar/EnvControls
